Simplify triggering by looking for command names instead of key combos
The BindingEvent generated when a keyboard shortcut is triggered
actually contains information on the command that was executed based on
the bindsym. This means that we don't have to deal with parsing of
key combos and synchronisation between Kitti3 and the config; instead
we can add a comment (arg) to i3's builtin `nop` command and search for
a match among incoming BindingEvents!