feat: realign Kitty when re-floated
The on_floated event handler is responsible for realignment of the
Kitty window when it is manually re-floated by the user after having
spent time in a tiled layout.

With the new event handler, the on_spawned function no longer needs to
call for alignment explicitly, as the act of floating the window will
trigger the handler.

Resolves #3