feat: manage any client
Introduce the -c, --client and -a,--cattr arguments, which allow Kitti3
to manage clients other than Kitty. -a selects the criterium attribute
used to locate a client instance (e.g. "title", "app_id"), and is only
required if -c is a custom command expression (e.g. "alacritty
--name {}"). Aternatively, -c can be a shorthand for a "known client"
which Kitti3 knows the appropriate expression and attribute for.
If neither options are provided, Kitti3 falls back on familiar behaviour
and spawns Kitty. If -a is provided but not -c, spawning is disabled
and it is assumed that spawning will be handled by the user. This
is handy for e.g. targeting the title of a specific browser window.

I'll make sure to litter the readme with examples and make a screencast
in the near future, as Kitti3 is turning into a bit of a swiss army
knife...