refactor: separate components into modules
Have been single-sourcing Kitti3 for a while to keep the "copy main.py
somewhere" installation approach viable, but this no longer makes sense
given the features we're accumulating.

Need to update the readme for the next release.

BREAKING CHANGE: It is no longer possible to "install" Kitti3 by
downloading main.py and making it executable.