feat(sway): slide-out animation
Enable support for slide-out animation on hide. Only triggered if the
client window's geometry meets its shape/position prediction. This
condition might be violated if the workspace's outer gaps have changed
since the last alignment, however, there doesn't appear to be a sensible
way to account for changing gaps in sway, and this is not likely to be a
frequent issue in practice.

The options --no-anim-show and --no-anim-hide provide control over which
animations are active.

Next up: pong.

BREAKING CHANGE: --anim-enter renamed --anim-show